Detach Player UI from InGameUI
