+++ DEFCON Z/Assets/Scripts/UI/PlayerUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6395cfcf4c907ab41ab75d2cc5b357a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
